created ArgoLiveRun class and changed structure
