+++ metaflow/plugins/argo/__init__.py
@@ -1,1 +1,1 @@
-from .live_run import trigger_run
+from .argo_live_run import trigger_argo_run
